agregado libros de prueba de leire

+++ db/1-create.sql
@@ -44,7 +44,7 @@ create table if not exists libro (
   id int unsigned auto_increment primary key,
   autor varchar(100) not null,
   serie varchar(30),
-  serie_num smallint(4),
+  serie_num float(4, 1),
   fecha_pub date not null,
   cantidad_reviews int unsigned not null default 0,
   nota_media float(3, 2) unsigned not null default 0,
